Refactored to move out nearby adapter from the activity
